% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-info.R
\name{get_kucoin_balances}
\alias{get_kucoin_balances}
\title{Get user's balance(s) list}
\usage{
get_kucoin_balances()
}
\value{
A \code{tibble} containing balance details
}
\description{
Get user's balance(s) list
}
\examples{

\dontrun{

# to run this example, make sure
# you already setup the API key
# in a proper .Renviron file

# import library
library(rucoin)

# get user's balance details
balances <- get_kucoin_balances()

# quick check
balances

}

}
