% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market-data.R
\name{get_kucoin_prices}
\alias{get_kucoin_prices}
\title{Get current KuCoin API server time}
\usage{
get_kucoin_prices(symbols, from, to, frequency)
}
\arguments{
\item{symbols}{A \code{character} vector of one or more pair symbol.}

\item{from}{A \code{date} or \code{datetime} object as a start of datetime range.}

\item{to}{A \code{date} or \code{datetime} object as an end of datetime range.}

\item{frequency}{A \code{character} vector of one which specify
the frequency option.}
}
\value{
A \code{tibble} containing prices data
}
\description{
Get current KuCoin API server time
}
\details{
List of supported frequencies:
\itemize{
\item \code{"1 minute"}
\item \code{"3 minutes"}
\item \code{"5 minutes"}
\item \code{"15 minutes"}
\item \code{"30 minutes"}
\item \code{"1 hour"}
\item \code{"2 hour"}
\item \code{"4 hour"}
\item \code{"6 hour"}
\item \code{"8 hour"}
\item \code{"12 hour"}
\item \code{"1 day"}
\item \code{"1 week"}
}
}
\examples{

# import library
library(rucoin)

# get one pair of symbol prices
prices <- get_kucoin_prices(
  symbols = "KCS/USDT",
  from = ymd_hms("2019-06-01 00:00:00"),
  to = ymd_hms("2019-06-02 00:00:00"),
  frequency = "1 hour"
)

prices

# get multiple pair of symbols prices
prices <- get_kucoin_prices(
  symbols = c("KCS/USDT", "BTC/USDT", "KCS/BTC"),
  from = ymd_hms("2019-06-01 00:00:00"),
  to = ymd_hms("2019-06-02 00:00:00"),
  frequency = "1 hour"
)

prices

}
